Ispravljene greške i dodani try catch blokovi; dodane slike iz postman-a

+++ MS-upravljanje/MS_upravljnaje/src/main/java/com/example/demo/Services/AktivnostService.java
@@ -2,11 +2,15 @@ package com.example.demo.Services;
 
 import java.util.Optional;
 
+import javax.persistence.RollbackException;
+import javax.validation.ConstraintViolationException;
+
 import com.example.demo.Entities.Aktivnost;
 import com.example.demo.Repositories.AktivnostRepositroy;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.TransactionSystemException;
 
 @Service
 public class AktivnostService {
@@ -24,16 +28,21 @@ public class AktivnostService {
     }
 
     public String addAktivnost(Aktivnost a){
-
         try {
             aktivnostrepository.save(a);
-        } catch (Exception e) {
-            return e.toString();
+        }
+        catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return e.getMessage();
+        } 
+        catch (Exception e) {
+            return e.getMessage();
         }
         return "Aktivnost saved";
     }
 
-    public String updateAktivnost(int id, Aktivnost a){
+    public String updateAktivnost(int id, Aktivnost a) {
+        if(!aktivnostrepository.findById(id).isPresent()) return "Aktivnost does not exist";
         try {
             aktivnostrepository.findById(id).map(aktivnost -> {
                 aktivnost.setMjesec(a.getMjesec());
@@ -41,13 +50,19 @@ public class AktivnostService {
                 aktivnost.setAktivnost(a.getAktivnost());
                 return aktivnostrepository.save(aktivnost);
             });
-        } catch(Exception e) {
+        }
+        catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return "Update Aktinvost error: " + e.getMessage();
+        }  
+        catch(Exception e) {
             return "Update Aktinvost error: " + e.toString();
         }
         return "Aktivnost successfully updated";
     }
 
     public String deleteAktivnost(int id){
+        if(!aktivnostrepository.findById(id).isPresent()) return "Aktivnost does not exist";
         try {
             aktivnostrepository.deleteById(id);
         } catch(Exception e) {
