+++ MS-upravljanje/MS_upravljnaje/src/main/java/com/example/demo/Services/RojenjeService.java
@@ -3,16 +3,21 @@ package com.example.demo.Services;
 import java.util.Optional;
 
 import com.example.demo.Entities.Rojenje;
+import com.example.demo.Repositories.KosnicaRepository;
 import com.example.demo.Repositories.RojenjeRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.TransactionSystemException;
 
 @Service
 public class RojenjeService {
 
     @Autowired
     RojenjeRepository rojenjeRepository;
+
+    @Autowired
+    KosnicaRepository kosnicaRepository;
 
     public Iterable<Rojenje> findAll() {
         return rojenjeRepository.findAll();
@@ -22,9 +27,17 @@ public class RojenjeService {
         return rojenjeRepository.findById(id);
     }
 
-    public String addRojenje(Rojenje r) {
+    public String addRojenje(Rojenje r, int idk) {
+        if(!kosnicaRepository.findById(idk).isPresent()) return "Kosnica does not exist";
         try {
-            rojenjeRepository.save(r);
+            kosnicaRepository.findById(idk).map(kosnica -> {
+                r.setKosnice(kosnica);
+                return rojenjeRepository.save(r);
+            });
+            
+        } catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return e.getMessage();
         } catch (Exception e) {
             return e.toString();
         }
@@ -32,6 +45,7 @@ public class RojenjeService {
     }
 
     public String updateRojenje(int id, Rojenje r){
+        if(!rojenjeRepository.findById(id).isPresent()) return "Rojenje does not exist";
         try {
             rojenjeRepository.findById(id).map(rojenje -> {
                 rojenje.setBrojmaticnjaka(r.getBrojmaticnjaka());
@@ -40,6 +54,9 @@ public class RojenjeService {
                 rojenje.setTipmaticnjaka(r.getTipmaticnjaka());
                 return rojenjeRepository.save(rojenje);
             });
+        } catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return "Update error: " + e.getMessage();
         } catch (Exception e) {
             return "Update error: " + e.toString();
         }
@@ -47,6 +64,7 @@ public class RojenjeService {
     }
 
     public String deleteRojenje(int id){
+        if(!rojenjeRepository.findById(id).isPresent()) return "Rojenje does not exist";
         try{
             rojenjeRepository.deleteById(id);;
         }
