+++ MS-upravljanje/MS_upravljnaje/src/main/java/com/example/demo/Services/SelidbaService.java
@@ -7,6 +7,7 @@ import com.example.demo.Repositories.SelidbaRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.TransactionSystemException;
 
 @Service
 public class SelidbaService {
@@ -25,13 +26,17 @@ public class SelidbaService {
     public String addSelidba(Selidba s) {
         try {
             selidbaRepository.save(s);
-        } catch (Exception e) {
+        } catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return e.getMessage();
+        }catch (Exception e) {
             return e.toString();
         }
         return "Selidba saved";
     }
 
     public String updateSelidba(int id, Selidba s){
+        if(!selidbaRepository.findById(id).isPresent()) return "Selidba does not exist";
         try{
             selidbaRepository.findById(id).map(selidba -> {
                 selidba.setBrojkosnica(s.getBrojkosnica());
@@ -41,13 +46,17 @@ public class SelidbaService {
                 selidba.setLokacija(s.getLokacija());
                 return selidbaRepository.save(selidba);
             }); 
-        } catch(Exception e) {
+        } catch (TransactionSystemException  ex) {
+            Throwable e = ex.getRootCause();
+            return "Update Selidba error: " + e.getMessage();
+        }catch(Exception e) {
             return "Update Selidba error: " + e.toString();
         }
         return "Selidba successfully updated";
     }
 
     public String delteSelidba(int id){
+        if(!selidbaRepository.findById(id).isPresent()) return "Selidba does not exist";
         try{
             selidbaRepository.deleteById(id);
         } catch(Exception e){
